merge cm changes to 24/12/15 misc: lis3dh_mot: Make sure ODR is always updated
Configuration changes are only possible when the sensor is enabled.
The kernel keeps internally an updated configuration of the sensor, but
this is only used when the sensor is completely turned off, which
might never happen. As consequence, some configuration changes might be
lost if users apply them when the sensor is disabled.

Make sure this never happens by applying the configuration changes
(currently REG1 only) users requested as soon as the sensor is enabled.

Change-Id: Id292024dcc12666ab5c2fcaa4789491fdeb59e2a

misc: lis3dh_mot: Refactor SET_MODE IOCTL

Switch to a bitmask for the sensor operating mode. This will allow
to handle composite modes in a simpler way. Also, simplify the code
that takes care of switching from one mode to another.

Change-Id: I5ed06b344dbbac119df3438c2cc0c2ce5a0adfe8

misc: lis3dh_mot: Fix CTRL_REG4 configuration when enabling IRQ

BLK_UPDATE is equal to 0x80, so the content of CTRL_REG4 is always set
to either 0x80 or 0x00 when the IRQ is enabled. Most likely the purpose
of the code was setting the threshold to G_2G while making that BDU was
set to 1, so change it to make this happen.

In addition to that, the value of resume_state[3] was set equal to
the content of CTRL_REG5 when it actually should contains the content
of CTRL_REG4. So move the assignement so that resume_state[3] contains
the correct value.

Change-Id: Icca6db20a990608da9f3c6ef3ee50acd0d171ff9

misc: lis3dh_mot: Enable IRQ and configure sensor separately

Use separate functions to enable the IRQ and configure the sensor.
This will allow to handle different modes in a better way.

Change-Id: I95e9ebb7f6e1ba39c7aec4eac610ab1f95a4cd36

misc: lis3dh_mot: Add movement detection mode

This mode allows to detect device movements. The sensor is a wakeup
source when working in this mode and never suspended.

This mode can't be enabled all together with the rotate mode.

Change-Id: Ia82e1aaac925b910d7567f7fbfa91220dd0b5b67

misc: lis3dh_mot: Report timestamp associated to event packet

The Android CDD strongly recommendeds to report the time the event
happened synchronized with the SystemClock.elapsedRealTimeNano() clock.

Change-Id: I3f0a8775bf096fe65e93ef61644e3f01bbac97d2